requirements.txt
aiogram==3.0.0b7
aiohttp>=3.8
python-dotenv>=0.20.0
